Add 'schema_name' parameter
This adds a parameter to the base class called `schema_name`, which can
be used to set a value for the `schema` property in a core's
`core.properties` file.

Prior to this commit, the module hardcoded a value for this based on the
version. For Solr >= 5.6.0, it insisted on using 'manage-schema';
for earlier versions, it used the classic `schema.xml`.

It may be desirable to use the classic `schema.xml` file-based schema
config rather than the managed config on current versions of Solr.

This is backwards-compatible. The value defaults to `undef` and a simple
case statement will use the original version-based default value if a
value is not specified.